리뷰 삭제 / 별점 구현

+++ src/main/java/com/project_250131/review/domain/Review.java
@@ -11,7 +11,7 @@ public class Review {
     private int storeId;
     private int userId;
     private String imagePath;
-    private Double point;
+    private int point;
     private String content;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
